Fix Windows build.
BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/10578039

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@11888 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/x64/lithium-codegen-x64.cc
@@ -945,9 +945,9 @@ void LCodeGen::DoMathFloorOfDiv(LMathFloorOfDiv* instr) {
         static_cast<double>(static_cast<uint64_t>(1) << shift) / divisor_abs;
     int64_t multiplier;
     if (multiplier_f - floor(multiplier_f) < 0.5) {
-        multiplier = floor(multiplier_f);
+        multiplier = static_cast<int64_t>(floor(multiplier_f));
     } else {
-        multiplier = floor(multiplier_f) + 1;
+        multiplier = static_cast<int64_t>(floor(multiplier_f)) + 1;
     }
     // The multiplier is a uint32.
     ASSERT(multiplier > 0 &&
